ESP32: Complete Build System Modernization and HAL Refactor
This commit consolidates the full ESP32 refactor, including:
- Data-driven build system infrastructure with chip capability database.
- Support for ESP32, S2, S3, C3, C6, and P4 variants.
- RISC-V toolchain support for C3/C6/P4.
- Modern ESP-IDF 5.5+ and 6.0 compatibility.
- Migration of core drivers (I2C, SPI, RMT, TWAI) to new callback APIs.
- Improved DroneCAN support with DNA server conflict avoidance.
- Implementation of MAVLink PARAM_EXT protocol for DroneCAN bridge.
- Core dump to flash and startup crash reporting.
- Enhanced diagnostics, memory protection, and hardenings.
- Cleanup of cross-platform compatibility issues (struct packing, math guards).
- Documentation updates including board migration guide.

Relates to https://github.com/ArduPilot/ardupilot/issues/30709

+++ libraries/AP_HAL/utility/DataRateLimit.cpp
@@ -18,7 +18,7 @@ uint32_t DataRateLimit::max_bytes(const float bytes_per_sec)
 
     // Get integer number of bytes and store the remainder
     float max_bytes_int;
-    remainder = modf(max_bytes, &max_bytes_int);
+    remainder = modff(max_bytes, &max_bytes_int);
 
     // Add 0.5 to make sure the float rounds to the correct int
     return uint32_t(max_bytes_int + 0.5);
